Preserve the Room3/Paging3 migration while resolving current master conflicts.

+++ app/src/main/java/one/mixin/android/api/response/web3/StakeAccount.kt
@@ -86,7 +86,7 @@ data class Lockup(
 @Parcelize
 data class Stake(
     @SerializedName("creditsObserved")
-    val creditsObserved: Int,
+    val creditsObserved: BigInteger,
     @SerializedName("delegation")
     val delegation: Delegation
 ) : Parcelable
